+++ HISTORY.rst
@@ -6,6 +6,13 @@ History
 0.0.6 (2019-02-27)
 ++++++++++++++++++
 
+* added a tokenize method to TeiReader
+* NE information written as rs-tags into TEI
+* minor things
+
+0.0.6 (2019-02-27)
+++++++++++++++++++
+
 * added langid to install_requires
 * new function `recogito.recogito_dump_to_spacy_ner` added
 * minor things